pyproject.toml
[project]
name = "langchain-masterclass"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.5.1",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.16",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.10",
    "onnxruntime<1.20",
    "pydantic>=2.12.5",
    "pypdf>=6.7.4",
    "python-dotenv>=1.2.1",
]
